Extract imgur parameters to parameter object
